\name{create_gist}
\alias{create_gist}
\title{Function that takes a list of files and creates payload for API}
\usage{
  create_gist(filenames, description = "", public = TRUE)
}
\arguments{
  \item{filenames}{names of files to post}

  \item{description}{brief description of gist (optional)}

  \item{public}{whether gist is public (defaults to TRUE)}
}
\description{
  Function that takes a list of files and creates payload
  for API
}

